implemented TokenFilter
